requirements.txt
dj-database-url == 0.4.1
django == 1.11.4
django-debug-toolbar == 1.6
django-dotenv == 1.4.1
django-extensions == 1.8.1
django-redis-cache == 1.7.1
djangorestframework == 3.6.3
drf-generators == 0.3.0
drf-nested-routers == 0.11.1
mysqlclient == 1.3.10
redis == 2.10.5
werkzeug == 0.11.15
